optimize config
